Refactor: App.tsx 위치 이동
