+++ src/app/style/index.css
@@ -1,0 +1,4 @@
+html, body {
+  background: #fff;
+  color: #000;
+}
